+++ config/splat.us.dra.yaml
@@ -272,6 +272,7 @@ segments:
       - [0x843B0, c, 843B0]
       - [0x84B88, c, 84B88]
       - [0x8A0A4, c, 8A0A4]
+      - [0x8BEF8, c, 8BEF8]
       - [0x8D3E8, c, 8D3E8]
       - [0x91EBC, c, 91EBC]
       - [0x92F60, c, 92F60]
